+++ bulk_index_request_easyjson.go
@@ -78,7 +78,7 @@ func easyjson9de0fcbfDecodeGithubComOlivereElastic(in *jlexer.Lexer, out *bulkIn
 				if out.IfSeqNo == nil {
 					out.IfSeqNo = new(int64)
 				}
-				*out.IfSeqNo = int(in.Int64())
+				*out.IfSeqNo = int64(in.Int64())
 			}
 		case "if_primary_term":
 			if in.IsNull() {
@@ -88,7 +88,7 @@ func easyjson9de0fcbfDecodeGithubComOlivereElastic(in *jlexer.Lexer, out *bulkIn
 				if out.IfPrimaryTerm == nil {
 					out.IfPrimaryTerm = new(int64)
 				}
-				*out.IfPrimaryTerm = int(in.Int64())
+				*out.IfPrimaryTerm = int64(in.Int64())
 			}
 		default:
 			in.SkipRecursive()
