+++ bulk_update_request_easyjson.go
@@ -68,7 +68,7 @@ func easyjson1ed00e60DecodeGithubComOlivereElastic(in *jlexer.Lexer, out *bulkUp
 				if out.IfSeqNo == nil {
 					out.IfSeqNo = new(int64)
 				}
-				*out.IfSeqNo = int(in.Int64())
+				*out.IfSeqNo = int64(in.Int64())
 			}
 		case "if_primary_term":
 			if in.IsNull() {
@@ -78,7 +78,7 @@ func easyjson1ed00e60DecodeGithubComOlivereElastic(in *jlexer.Lexer, out *bulkUp
 				if out.IfPrimaryTerm == nil {
 					out.IfPrimaryTerm = new(int64)
 				}
-				*out.IfPrimaryTerm = int(in.Int64())
+				*out.IfPrimaryTerm = int64(in.Int64())
 			}
 		default:
 			in.SkipRecursive()
